refactor(UserAccount): restructure exceptions and value objects into dedicated subdirectories
